Move box-drawing glyphs from Escritoire into `tessellate`
`BoxDrawing`, `BoxLine` and `LineCharset` are layout vocabulary, not
table vocabulary: relocating them lets Ultimatum's borders (and any
other module) draw correct junctions from the same 256-entry table
Escritoire uses. Pure motion; the `soundness` package surface is
unchanged.
