+++ lib/tessellate/src/core/tessellate.LineCharset.scala
@@ -30,37 +30,24 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package escritoire
+package tessellate
 
-import anticipation.*
-import proscenium.compat.*
-import gossamer.*
+enum LineCharset:
+  case Default, Rounded, Ascii
 
-object BoxDrawing:
-  val asciiChars: Array[Char]^{} =
-    List
-      ( t" -- |+++|++  + +--- +++ +++ +++ --- +++++++  + +   -+ ++   -+ ++|+++|+++|+++    +++ +++ ",
-        t"+++     ++++++++++++    + +++ +++ ++    |++ |+++|++  + ++++ +++ +++ +++ +++ +++++ +  + +",
-        t"   -+ ++   -+ ++ + +     + +|+|++++     +++ +++  + +     + +|+|++ ++    + +++ ++" )
+  // Each branch is ascribed: the inferred union of the three frozen vals picks up a
+  // fresh `any.rd` that cannot flow back into `^{}`.
+  def apply(): Array[Char]^{} = this match
+    case Default => BoxDrawing.defaultChars: Array[Char]^{}
+    case Rounded => BoxDrawing.roundedChars: Array[Char]^{}
+    case Ascii   => BoxDrawing.asciiChars: Array[Char]^{}
 
-    . join
-    . chars
 
-  val defaultChars: Array[Char]^{} =
-    List
-      ( t" ╴╸ ╷┐┑╕╻┒┓  ╖ ╗╶─╾ ┌┬┭ ┎┰┱ ╓╥╓ ╺╼━ ┍┮┯╕┏┲┳  ╖ ╗   ═╒ ╒╤   ═╔ ╔╦╵┘┙╛│┤┥╡╽┧┪╛    └┴┵ ├┼┽ ",
-        t"┟╁╅     ┕┶┷╛┝┾┿╡┢╆╈╛    ╘ ╘╧╞ ╞╪╘ ╘╧    ╹┚┛ ╿┦┩╕┃┨┫  ╖ ╗┖┸┹ ┞╀╃ ┠╂╊ ╓╥╓ ┗┺┻ ┡╄╇╕┣ ╋  ╖ ╗",
-        t"   ═╒ ╒╤   ═╔ ╔╦ ╜ ╝     ╜ ╝║╢║╣╙╨╙     ╙╨╙ ╟╫╟  ╜ ╝     ╜ ╝║╢║╣╚ ╚╩    ╚ ╚╩╠ ╠╬" )
+  def apply
+    ( top:    BoxLine = BoxLine.Blank,
+      right:  BoxLine = BoxLine.Blank,
+      bottom: BoxLine = BoxLine.Blank,
+      left:   BoxLine = BoxLine.Blank )
+  :   Char =
 
-    . join
-    . chars
-
-  val roundedChars: Array[Char]^{} = defaultChars.map:
-    case '┌'  => '╭'
-    case '┘'  => '╯'
-    case '┐'  => '╮'
-    case '└'  => '╰'
-    case char => char
-
-  def simple(vertical: BoxLine, horizontal: BoxLine, charset: LineCharset): Char =
-    charset(vertical, horizontal, vertical, horizontal)
+    this().readable(left.ordinal + bottom.ordinal*4 + right.ordinal*16 + top.ordinal*64)
